Module algeo::core::num

Traits

EpsilonEquality

Field

One

StabilityCmp
Partial order to indicate which element has greater stability. More precisely, it is used to determine what number is best to divide by. For lu decomposition (or just Gaussian elimination in general), it is always best to divide by large floats, since this will lead to the least rounding/ float point problems.

Zero