Trait algeo::core::num::StabilityCmp

pub trait StabilityCmp {
    fn stability_cmp(&self, other: &Self) -> Option<Ordering>;
}

Partial order to indicate which element has greater stability. More precisely, it is used to determine what number is best to divide by. For lu decomposition (or just Gaussian elimination in general), it is always best to divide by large floats, since this will lead to the least rounding/ float point problems.

Needs to satisfy 0<= every number, (and every number !<= 0) in order to avoid divide by 0.

Required methods

fn stability_cmp(&self, other: &Self) -> Option<Ordering>

Loading content...