Struct algeo::poly::ring::Polynomial

pub struct Polynomial<F: Field> {
    pub terms: Vec<Term<F>>,
}

a polynomial with coefficients in the field F

Fields
terms: Vec<Term<F>>
Implementations

impl<F: Field> Polynomial<F>

pub fn new_unchecked(terms: Vec<Term<F>>) -> Self

Returns the polynomial with terms exactly vec.

This does no checking for zero terms

pub fn new(terms: Vec<Term<F>>) -> Self

Returns polynomial with terms from vec, filtered for zero coefficients

Ideally, this should not be necessary; polynomial computations should be careful to keep themselves clean operations should be structured such that no additional filtering is necessary except during creation

pub fn zero() -> Self

Returns the zero polynomial.

Currently, this is a polynomial with no terms

pub fn is_zero(&self) -> bool

should not be used until we can guarantee that zero terms get filtered

pub fn monomial(t: Term<F>) -> Self

returns a polynomial with only the term t

pub fn constant(coef: F) -> Self

pub fn one() -> Self

returns the polynomial with only the constant term 1

pub fn terms(&self) -> Iter<'_, Term<F>>

Returns an iterator over (immutably) borrowed terms: Item = &Term<F>.

Should replace return with in-house iterator struct.

pub fn terms_mut(&mut self) -> IterMut<'_, Term<F>>

Returns an iterator over mutably borrowed terms: Item = &mut Term<F>.

Should replace return with in-house iterator struct.

pub fn eval(&self, x: &[F]) -> F

Evaluates self in F[x_1, ..., x_n] at the point x in F^m.

Panics

Method panics if m < n, i.e., if x does not provide values up the maximum index variable of self.